requirements.txt
playwright>=1.58.0
pytest-playwright>=0.7.2
pytest-xdist>=3.8.0
pytest-html>=4.2.0
python-dotenv>=1.2.2
requests>=2.32.5
requests-oauthlib>=2.0.0